Related to #10425 Updated XML-doc comments around CodeDomComponentSerializationService
