+++ src/System.Windows.Forms.Design/src/System/ComponentModel/Design/Serialization/CodeDomComponentSerializationService.CodeDomSerializationStore.ComponentListCodeDomSerializer.cs
@@ -72,7 +72,7 @@ public sealed partial class CodeDomComponentSerializationService
                     if (objectState.TryGetValue(name, out CodeDomComponentSerializationState? state))
                     {
                         PopulateCompleteStatements(state.Code, name, completeStatements, _expressions);
-                        PopulateCompleteStatements(state.Ctx, name, completeStatements, _expressions);
+                        PopulateCompleteStatements(state.Context, name, completeStatements, _expressions);
                     }
                 }
